requirements.txt
# Core ML and DL libraries
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
opencv-python==4.8.1.78
Pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2

# MLOps and Experiment Tracking
mlflow>=1.28.0
mlflow-skinny>=1.28.0
dvc==3.22.0
dvc[s3]==3.22.0

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2

# Data Processing
tqdm==4.66.1
PyYAML==6.0.1

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Code Quality
black==23.9.1
flake8==6.1.0
mypy==1.6.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
gunicorn==21.2.0

# Monitoring and Logging
psutil==5.9.6
evidently==0.4.6

# Deployment
docker==6.1.3
